Social fix

+++ social/delete_post.php
@@ -3,9 +3,12 @@ include('db.php');
 
 $id_sosmed = $_GET['codigo'];
 
-$sql = "DELETE FROM Sosmed WHERE id_sosmed = :id_sosmed";
-$stmt = $conn->prepare($sql);
-$stmt->execute(['id_sosmesd' => $id_sosmed]);
+if ($id_sosmed) {
+    $sql = "DELETE FROM Sosmed WHERE id_sosmed = :id_sosmed";
+    $stmt = $conn->prepare($sql);
+    $stmt->bindParam(':id_sosmed', $id_sosmed, PDO::PARAM_INT);
+    $stmt->execute();
+}
 
-header("Location: ../views/index.php");
+header("Location: social.php");
 ?>
